added instructions + cleanup
