+++ examples/ssx-test-serverless-dynamodb-api/package.json
